changed to float

+++ methods/parallel_cpp/kernel.cu
@@ -124,22 +124,22 @@ void free_model(DeviceModel* device_model) {
     free_buffer(&device_model->wpe);
     free_buffer(&device_model->lm_head);
 
-    for (DeviceBuffer<double>& buffer : device_model->attn_wq) {
+    for (DeviceBuffer<float>& buffer : device_model->attn_wq) {
         free_buffer(&buffer);
     }
-    for (DeviceBuffer<double>& buffer : device_model->attn_wk) {
+    for (DeviceBuffer<float>& buffer : device_model->attn_wk) {
         free_buffer(&buffer);
     }
-    for (DeviceBuffer<double>& buffer : device_model->attn_wv) {
+    for (DeviceBuffer<float>& buffer : device_model->attn_wv) {
         free_buffer(&buffer);
     }
-    for (DeviceBuffer<double>& buffer : device_model->attn_wo) {
+    for (DeviceBuffer<float>& buffer : device_model->attn_wo) {
         free_buffer(&buffer);
     }
-    for (DeviceBuffer<double>& buffer : device_model->mlp_fc1) {
+    for (DeviceBuffer<float>& buffer : device_model->mlp_fc1) {
         free_buffer(&buffer);
     }
-    for (DeviceBuffer<double>& buffer : device_model->mlp_fc2) {
+    for (DeviceBuffer<float>& buffer : device_model->mlp_fc2) {
         free_buffer(&buffer);
     }
 
@@ -207,9 +207,9 @@ void free_workspace(DeviceWorkspace* workspace) {
 
 __global__ void embedding_lookup_kernel(
     const int* tokens,
-    const double* wte,
-    const double* wpe,
-    double* embeddings,
+    const float* wte,
+    const float* wpe,
+    float* embeddings,
     int batch_size,
     int batch_seq_length,
     int usable_seq_len,
@@ -228,15 +228,15 @@ __global__ void embedding_lookup_kernel(
     const int token_idx = batch_idx * batch_seq_length + pos;
     const int token_id = tokens[token_idx];
 
-    const double token_val = wte[token_id * n_embd + col];
-    const double pos_val = wpe[pos * n_embd + col];
+    const float token_val = wte[token_id * n_embd + col];
+    const float pos_val = wpe[pos * n_embd + col];
     embeddings[idx] = token_val + pos_val;
 }
 
 __global__ void add_vec_kernel(
-    const double* left,
-    const double* right,
-    double* output,
+    const float* left,
+    const float* right,
+    float* output,
     const int n
 ) {
     int tid = blockDim.x * blockIdx.x + threadIdx.x;
@@ -246,8 +246,8 @@ __global__ void add_vec_kernel(
 }
 
 __global__ void rmsnorm_kernel(
-    const double* input,
-    double* output,
+    const float* input,
+    float* output,
     int n_embd,
     int useable_seq_len, 
     int num_batches
@@ -259,16 +259,16 @@ __global__ void rmsnorm_kernel(
 
     int start = idx * n_embd;
 
-    double mean_square = 0.0;
+    float mean_square = 0.0f;
 
     for (int i = 0; i < n_embd; i++) {
-        double value = input[start + i];
+        float value = input[start + i];
         mean_square += value * value;
     }
 
-    mean_square /= (double)n_embd;
+    mean_square /= static_cast<float>(n_embd);
 
-    double scale = 1.0 / sqrt(mean_square + 1e-5);
+    float scale = 1.0f / sqrtf(mean_square + 1e-5f);
 
     for (int i = 0; i < n_embd; i++) {
         output[start + i] = input[start + i] * scale;
@@ -277,20 +277,20 @@ __global__ void rmsnorm_kernel(
 
 
 __global__ void linear_tiled_kernel(
-    const double* input,
-    const double* weights,
-    double* output,
+    const float* input,
+    const float* weights,
+    float* output,
     int total_tokens,
     int out_dim,
     int in_dim
 ) {
-    __shared__ double input_tile[kLinearTile][kLinearTile];
-    __shared__ double weight_tile[kLinearTile][kLinearTile];
+    __shared__ float input_tile[kLinearTile][kLinearTile];
+    __shared__ float weight_tile[kLinearTile][kLinearTile];
 
     const int row = blockIdx.x * kLinearTile + threadIdx.x;
     const int col = blockIdx.y * kLinearTile + threadIdx.y;
 
-    double sum = 0.0;
+    float sum = 0.0f;
     const int tile_count = (in_dim + kLinearTile - 1) / kLinearTile;
 
     for (int tile_idx = 0; tile_idx < tile_count; ++tile_idx) {
@@ -298,9 +298,9 @@ __global__ void linear_tiled_kernel(
         const int weight_col = tile_idx * kLinearTile + threadIdx.x;
 
         input_tile[threadIdx.x][threadIdx.y] =
-            (row < total_tokens && input_col < in_dim) ? input[row * in_dim + input_col] : 0.0;
+            (row < total_tokens && input_col < in_dim) ? input[row * in_dim + input_col] : 0.0f;
         weight_tile[threadIdx.y][threadIdx.x] =
-            (col < out_dim && weight_col < in_dim) ? weights[col * in_dim + weight_col] : 0.0;
+            (col < out_dim && weight_col < in_dim) ? weights[col * in_dim + weight_col] : 0.0f;
 
         __syncthreads();
 
@@ -319,7 +319,7 @@ __global__ void linear_tiled_kernel(
 
 
 __global__ void relu_kernel(
-    double* input,
+    float* input,
     int n
 ) {
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
@@ -328,14 +328,14 @@ __global__ void relu_kernel(
 
     if (input[idx] >= 0) return; 
 
-    input[idx] = 0.0; 
+    input[idx] = 0.0f; 
 }
 
 __global__ void self_attn_kernel(
-    const double* q,
-    const double* k_layer,
-    const double* v_layer,
-    double* attn_out,
+    const float* q,
+    const float* k_layer,
+    const float* v_layer,
+    float* attn_out,
     int batch_size,
     int usable_seq_len,
     int n_head,
@@ -356,43 +356,43 @@ __global__ void self_attn_kernel(
     const int token_start = sequence_start + token_pos * n_embd;
     const int head = col / head_dim;
     const int head_start = head * head_dim;
-    const double scale = 1.0 / sqrt(static_cast<double>(head_dim));
+    const float scale = 1.0f / sqrtf(static_cast<float>(head_dim));
 
     if (head >= n_head) {
         return;
     }
 
-    double max_logit = -INFINITY;
+    float max_logit = -INFINITY;
     for (int t = 0; t <= token_pos; ++t) {
         const int past_start = sequence_start + t * n_embd;
-        double dot = 0.0;
+        float dot = 0.0f;
         for (int j = 0; j < head_dim; ++j) {
             dot += q[token_start + head_start + j] * k_layer[past_start + head_start + j];
         }
-        const double score = dot * scale;
+        const float score = dot * scale;
         if (score > max_logit) {
             max_logit = score;
         }
     }
 
-    double exp_sum = 0.0;
+    float exp_sum = 0.0f;
     for (int t = 0; t <= token_pos; ++t) {
         const int past_start = sequence_start + t * n_embd;
-        double dot = 0.0;
+        float dot = 0.0f;
         for (int j = 0; j < head_dim; ++j) {
             dot += q[token_start + head_start + j] * k_layer[past_start + head_start + j];
         }
-        exp_sum += exp(dot * scale - max_logit);
+        exp_sum += expf(dot * scale - max_logit);
     }
 
-    double weighted_value = 0.0;
+    float weighted_value = 0.0f;
     for (int t = 0; t <= token_pos; ++t) {
         const int past_start = sequence_start + t * n_embd;
-        double dot = 0.0;
+        float dot = 0.0f;
         for (int j = 0; j < head_dim; ++j) {
             dot += q[token_start + head_start + j] * k_layer[past_start + head_start + j];
         }
-        const double weight = exp(dot * scale - max_logit) / exp_sum;
+        const float weight = expf(dot * scale - max_logit) / exp_sum;
         weighted_value += weight * v_layer[past_start + col];
     }
 
@@ -401,8 +401,8 @@ __global__ void self_attn_kernel(
 
 
 __global__ void cross_entropy_loss_kernel(
-    const double* logits,
-    double* loss,
+    const float* logits,
+    float* loss,
     const int* tokens,
     int batch_size,
     int batch_seq_length,
@@ -413,27 +413,27 @@ __global__ void cross_entropy_loss_kernel(
         return;
     }
 
-    double total_loss = 0.0;
+    float total_loss = 0.0f;
     for (int batch_idx = 0; batch_idx < batch_size; ++batch_idx) {
         for (int pos = 0; pos < usable_seq_len; ++pos) {
             const int logits_start = (batch_idx * usable_seq_len + pos) * vocab_size;
             const int target_token = tokens[batch_idx * batch_seq_length + pos + 1];
-            double max_logit = logits[logits_start];
+            float max_logit = logits[logits_start];
             for (int vocab_idx = 1; vocab_idx < vocab_size; ++vocab_idx) {
-                const double value = logits[logits_start + vocab_idx];
+                const float value = logits[logits_start + vocab_idx];
                 if (value > max_logit) {
                     max_logit = value;
                 }
             }
 
-            double exp_sum = 0.0;
+            float exp_sum = 0.0f;
             for (int vocab_idx = 0; vocab_idx < vocab_size; ++vocab_idx) {
-                exp_sum += exp(logits[logits_start + vocab_idx] - max_logit);
+                exp_sum += expf(logits[logits_start + vocab_idx] - max_logit);
             }
-            total_loss += log(exp_sum) + max_logit - logits[logits_start + target_token];
+            total_loss += logf(exp_sum) + max_logit - logits[logits_start + target_token];
         }
     }
-    *loss = total_loss / static_cast<double>(batch_size * usable_seq_len);
+    *loss = total_loss / static_cast<float>(batch_size * usable_seq_len);
 }
 
 void launch_embedding(const DeviceModel& device_model, DeviceWorkspace* workspace, const ModelConfig& config, const BatchTokens& batch) {
@@ -457,8 +457,8 @@ void launch_embedding(const DeviceModel& device_model, DeviceWorkspace* workspac
 
 
 void launch_rmsnorm(
-    const double* input,
-    double* output,
+    const float* input,
+    float* output,
     int n_embd,
     int batch_size,
     int usable_seq_len
@@ -481,9 +481,9 @@ void launch_rmsnorm(
 
 
 void launch_linear(
-    const double* input,
-    double* output,
-    const double* weights, 
+    const float* input,
+    float* output,
+    const float* weights, 
     int in_dim, 
     int out_dim, 
     int batch_size,
@@ -510,10 +510,10 @@ void launch_linear(
 
 
 void launch_self_attn(
-    const double* q,
-    const double* k_layer,
-    const double* v_layer,
-    double* attn_out,
+    const float* q,
+    const float* k_layer,
+    const float* v_layer,
+    float* attn_out,
     int batch_size,
     int usable_seq_len,
     int n_head,
@@ -543,9 +543,9 @@ void launch_self_attn(
 
 
 void launch_vec_add(
-    const double* left,
-    const double* right,
-    double* output,
+    const float* left,
+    const float* right,
+    float* output,
     const int n
 ) {
     const auto launch = make_1d_launch(
@@ -564,7 +564,7 @@ void launch_vec_add(
 
 
 void launch_relu(
-    double* input, 
+    float* input, 
     int size
 ) {
     const auto launch = make_1d_launch(
@@ -612,8 +612,8 @@ void launch_transformer(const DeviceModel& device_model, DeviceWorkspace* worksp
         launch_linear(workspace->norm.ptr, workspace->q.ptr, device_model.attn_wq[layer_idx].ptr, config.n_embd, config.n_embd, batch.batch_size, usable_seq_len);
         
         //* Find the layer in the cache 
-        double* k_layer = workspace->k_cache.ptr + layer_idx * batch.batch_size * usable_seq_len * config.n_embd;
-        double* v_layer = workspace->v_cache.ptr + layer_idx * batch.batch_size * usable_seq_len * config.n_embd;
+        float* k_layer = workspace->k_cache.ptr + layer_idx * batch.batch_size * usable_seq_len * config.n_embd;
+        float* v_layer = workspace->v_cache.ptr + layer_idx * batch.batch_size * usable_seq_len * config.n_embd;
         
         launch_linear(workspace->norm.ptr, k_layer, device_model.attn_wk[layer_idx].ptr, config.n_embd, config.n_embd, batch.batch_size, usable_seq_len);
         launch_linear(workspace->norm.ptr, v_layer, device_model.attn_wv[layer_idx].ptr, config.n_embd, config.n_embd, batch.batch_size, usable_seq_len);
